requirements.txt
altair==5.5.0
asgiref==3.11.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
crispy-bootstrap5==2025.6
dj-database-url==3.1.2
Django==4.2.28
django-cors-headers==4.9.0
django-crispy-forms==2.5
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
gitdb==4.0.12
GitPython==3.1.45
gunicorn==25.1.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.12.0
numpy==2.3.5
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.1
psycopg2==2.9.11
psycopg2-binary==2.9.11
pyarrow==21.0.0
pydeck==0.9.1
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.29.0
six==1.17.0
smmap==5.0.2
sqlparse==0.5.5
streamlit==1.51.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
whitenoise==6.12.0
